// binotto-app/src/types/technician.ts
export type TechnicianProfileForm = {
  fullName: string;
  nickname: string;
  email: string;
  birthDate: string;
  nationality: string;
  secondNationality: string;
  primaryPhone: string;
  primaryPhoneCountryCode: string;
  primaryPhoneCountryIso: string;
  secondaryPhone: string;
  secondaryPhoneCountryCode: string;
  secondaryPhoneCountryIso: string;
  cpf: string;
  cnpj: string;
  companyFantasyName: string;
  companyLegalName: string;
  street: string;
  number: string;
  complement: string;
  city: string;
  state: string;
  zipCode: string;
  country: string;
  workCountries: string[];
};